feat(matching-C): carve RepairMenuItemDraw — DrawItemMenuLineLong param s8->int (permuter-found)
decomp-permuter on RepairMenuItemDraw (8/92) didn't reach 0 itself, but its best
candidate's source diff revealed the residual: type-narrowing (D89) — JP declares
DrawItemMenuLineLong's 3rd param int, not s8. Applied the JP-correct fix in
bmitem.h -> 0/92. No carved caller uses it; full make clean && make -j compare
byte-matches (shared-header verified). matching-C 83.31->83.33%. Permuter-as-diagnostic.

+++ include/bmitem.h
@@ -165,7 +165,7 @@ s8 CanUnitUseWeaponNow(struct Unit* unit, int item);
 s8 CanUnitUseStaff(struct Unit* unit, int item);
 s8 CanUnitUseStaffNow(struct Unit* unit, int item);
 void DrawItemMenuLine(struct Text* text, int item, s8 isGrayed, u16* mapOut);
-void DrawItemMenuLineLong(struct Text* text, int item, s8 isGrayed, u16* mapOut);
+void DrawItemMenuLineLong(struct Text* text, int item, int isGrayed, u16* mapOut);
 void DrawItemMenuLineNoColor(struct Text* text, int item, u16* mapOut);
 void DrawItemStatScreenLine(struct Text* text, int item, int nameColor, u16* mapOut);
 u16 GetItemAfterUse(int item);
